Fix TestRequestPoolService, do not use URL.getPath()
URL.getPath() will return a valid URL, which means that some unsafe ASCII
characters are encoded for URLs. In some test environments we had repos
with '@' in the path and this was not handled properly by this test when
we try to load test resources.

Change-Id: I74314c40d7c70b908456cd9263263e83c79e0664
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2108
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins
(cherry picked from commit a7118d509bb897086966ad98e95052176246801f)
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2115
Reviewed-by: Matthew Jacobs <mj@cloudera.com>