Changes to Avro C library to allow parsing decimal schemas
This augments the schema representation to handle decimals, storing
the scale and precision in the decimal's schema object. This is the
only functionality required by Impala. Elsewhere in the code decimals
are handled the same as bytes, which should mean that decimal values
look like bytes values, but I haven't tested this functionality.

Change-Id: Ie12d2d6b238b5fc67a7adc207e0fec7fc0e7683d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2436
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: jenkins