Remove dependency on Hive for Impala Avro schema parsing
This change removes Impala's dependency on Hive for parsing the Avro schema. To remove the
dependency I use the Avro library to parse the schema and then reused some of Hive's AvroSerDe
code to perform the Avro -> Hive type mapping.

It also adds support for parsing DECIMAL type information from an Avro schema.

Change-Id: I4359210ce50ddc3c0d03fe9eb30d35cc8e45a797
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2460
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins