CDH-19184: Impala should show impersonated user (if there is one) rather than connected user
Currently, we always display the 'User' as the connected user in the debug webpage and
runtime profiles. This is confusing when impersonation + authorization is enabled because
there is not an easy way to find the impersonated user other than looking at the audit
log records. This change does the following:
* Updates the "User" field in the runtime profile to show the "effective user".
  The effective user is the connected user if there is no impersonated user,
  otherwise it is the impersonated user. This should help CM display the correct user
  as well.
* Add two new fields in the runtime profile "Connected User" & "Impersonated User"
  to make it easier to tell which user is which.
* Update the /queries debug webpage to show the effective user rather than the
  connected user.

Change-Id: I639de6738242d2c378e785271a72257301a53ade
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2863
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins
(cherry picked from commit d4ad768780dfdfe0874f2b3e9c59074f1c3685d7)
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2935