[CDH5] CDH-19662: Add sentry policy refresh in 'invalidate metadata'
Updates 'invalidate metadata' to force an update of the Sentry Policy Service metadata,
if Sentry is enabled.
Also fixes an issue with broken connection to the Sentry Policy Service by creating
a new connection for each Sentry RPC. This is not desireable (we should be able to
cache the connections), but currently there is no way to detect broken connections
and "re-open" them (tracked by SENTRY-296).

I am working on adding a test for this, but since Impala does not yet support grant
revoke it is a bit tricky.

Change-Id: I681037beeec9cbf378126d799f6e07c89a4dc29c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3054
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins