IMPALA-1036: Exit gracefully for misconfigured scratch directories.
Log an error if a scratch directory does not exist, is not a directory,
is not writable etc.
Log a warning if there is less than 1GB (arbitrarily chosen) of
available space on a particular configured disk.
Move TmpFileMgr initialization from common runtime to ImpalaServer.
Also fixes bug where an invalid disk id is assigned to a WriteRange.
Similar to HdfsScanNode::AllocateScanRange, a random disk id is assigned.

Change-Id: I869c005485b0e7066cbe90f9d5cb20b431fdd025
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2961
Reviewed-by: Srinath Shankar <sshankar@cloudera.com>
Tested-by: jenkins
Reviewed-on: http://gerrit.ent.cloudera.com:8080/3052