IMPALA-1041: Detect partitions that contain invalid MD
Previously, we would load a partition but would not perform any validation on whether
the partition metadata was actually well formed (could be converted into our
internal thrift representation). This updates our table loading logic to do
add an additional validation check after creating a partition to ensure the metadata is
well formed.
The hang reported in IMPALA-1041 was because we were unable to convert the target table
toThrift() so it never got sent to the impalad from the catalog server due to invalid
partition metadata.

Change-Id: I4848427cbc923d6a0e515ba5154e900981dbf9ae
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2981
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: jenkins