IMPALA-1039: Handling a corner case in Parquet metadata expected rows validation.
IMPALA-1022 introduced a validation of the expected number of rows in Parquet metadata,
but it was not handling correctly the corner case where the expected number of rows is
a multiple of the row batch size. In that case, it was erroneously trying to write one
more row in the batch and was causing memory error (asan).
This batch handles this corner case, by not attempting to read one more additional row
when the scanner reaches both the end of expected number of rows and the end of the
row batch.

Change-Id: I45e8dc4f25bb4af6cd72c1e14e578a5e001957a1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/2989
Reviewed-by: Nong Li <nong@cloudera.com>
Tested-by: jenkins